requirements.txt
pip install -r requirements.txt 
django-crispy-forms